modularization of code
