+++ srcs/requirements/apache/volume/css/charts.css
@@ -5,6 +5,7 @@ body, html
     display: flex;
     justify-content: center;
     align-items: center;
+	background-color: rgb(198, 137, 255);
 }
 
 .grid-container 
@@ -23,12 +24,41 @@ body, html
     justify-content: center;
     align-items: center;
     height: 100%;
-    border: 1px solid #ccc;
+    border: 1px solid #1b1b1b;
 }
 
-.grid-item.navbar {
+.grid-item.navbar 
+{
     width: 100%;
     justify-content: flex-start; /* Aligne les éléments à gauche */
     align-items: flex-start; /* Aligne les éléments en haut */
 	border: none;
 }
+/* #app 
+{
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+    height: 100vh;
+    width : 100vw
+} */
+
+#grid-charts
+{
+    flex-grow: 1; 
+    justify-content: center;
+    display: grid;
+	grid-template-columns: 1fr 1fr;
+	grid-template-rows: 1fr 1fr;
+    height:90%;
+    width:90%;
+    border:none;
+}
+
+/* .grid-item
+{
+	border: 2px solid rgb(0,0,0);
+	width : 90%;
+	height : 90%;
+} */
