+++ srcs/requirements/apache/volume/js/utils/loadPage.js
@@ -1,0 +1,58 @@
+export function loadCss(page)
+{
+	if (page) 
+	{
+		const link = document.createElement('link');
+		link.rel = 'stylesheet';
+		link.href = `css/${page}.css`;
+		link.dataset.page = page;
+		document.head.appendChild(link);
+	}	
+}
+
+export async function loadHtml(page,div)
+{
+	const response = await fetch(`pages/${page}.html`);
+	if (!response.ok) {
+		throw new Error('Network response was not ok');
+	}
+	const html = await response.text();
+	const container = document.getElementById(div);
+	
+	if (container) {
+		container.innerHTML = html;
+	} 
+	else
+		console.error(`Le conteneur ${div} n'existe pas`);
+		
+}
+
+export async function loadScript(page) 
+{
+	const scriptId = `script-${page}`;
+	let existingScript = document.getElementById(scriptId);
+
+	if (existingScript) {
+	existingScript.remove();
+	}
+
+	// Créer un élément script
+	const script = document.createElement('script');
+	script.src = `js/${page}.js`;
+	script.id = scriptId;
+	script.async = true;
+	script.type = 'module';
+
+	// Attendre que le script soit chargé
+	return new Promise((resolve) => {
+		script.onload = () => {
+			console.log(`Script ${page}.js chargé avec succès.`);
+			resolve();
+		};
+		script.onerror = () => {
+			console.warn(`Le script ${page}.js n'a pas pu être chargé. Aucune erreur retournée.`); 
+			resolve();
+		};
+		document.body.appendChild(script);
+	});
+}
