+++ srcs/requirements/apache/volume/pages/charts.html
@@ -1,0 +1,13 @@
+
+<div id="charts-page">
+	<div class="grid-container">
+		<div class="grid-item navbar" id = "navbar"></div>
+        <div id ="grid-charts">
+            <div class="grid-item" id="bars-score-by-game"></div>
+            <div class="grid-item" id="pie-win-defeat"></div>
+            <div class="grid-item" id="other-graph"></div>
+            <div class="grid-item" id="other-graph2"></div>
+        </div>
+        
+	</div>
+</div>
