+++ srcs/requirements/apache/volume/pages/home.html
@@ -1,69 +1,34 @@
 <div id="main-page">
-    <div class="grid-container">
-        <div class="grid-item navbar">
-            <nav class="navbar navbar-expand-lg bg-body-tertiary">
-                <div class="container-fluid">
-                  <a class="navbar-brand" href="#">Transcendance</a>
-                  <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
-                    <span class="navbar-toggler-icon"></span>
-                  </button>
-                  <div class="collapse navbar-collapse" id="navbarSupportedContent">
-                    <ul class="navbar-nav me-auto mb-2 mb-lg-0">
-                      <li class="nav-item">
-                        <a class="nav-link active" id="home_button" aria-current="page" href="#">Home</a>
-                      </li>
-                      <li class="nav-item">
-                        <a class="nav-link" href="#">Link</a>
-                      </li>
-                      <li class="nav-item dropdown">
-                        <a class="nav-link dropdown-toggle" href="#" role="button" data-bs-toggle="dropdown" aria-expanded="false">
-                          Dropdown
-                        </a>
-                        <ul class="dropdown-menu">
-                          <li><a class="dropdown-item" href="#">Action</a></li>
-                          <li><a class="dropdown-item" href="#">Another action</a></li>
-                          <li><hr class="dropdown-divider"></li>
-                          <li><a class="dropdown-item" href="#">Something else here</a></li>
-                        </ul>
-                      </li>
-                    </ul>
-                    <div class="d-flex ms-auto">
-                        <a href=#login class="btn btn-outline-success me-2" id="button_to_login" div="main-grid">login</a>
-                        <a href=#register class="btn btn-outline-success" id="button_to_register" div="main-grid">register</a>
-							          <button class="btn btn-outline-success" id="button_to_42api">42-connector</button>
-                      </div>
-                  </div>
-                </div>
-            </nav>
-        </div>
-        <div class="grid-item" id = "main-grid">
-            
-            <div class="card square-card">
-                <!-- <img src="..." class="card-img-top" alt="..."> -->
-                <div class="card-body">
-                <h5 class="card-title">RolanGapong</h5>
-                <p class="card-text">RolanGapong est une version revisitée du célèbre jeu Pong, avec des thèmes inspirés des plus grands tournois de tennis.</p>
-                <a href="#pong" class="btn btn-primary" id="button_to_pong" div="app">Play</a>
-                </div>
-            </div>
-            <div class="card square-card">
-                <!-- <img src="..." class="card-img-top" alt="..."> -->
-                <div class="card-body">
-                <h5 class="card-title">Card Game</h5>
-                <p class="card-text">Card Game est un jeu de cartes classique où les joueurs doivent retrouver des paires identiques parmi des cartes disposées face cachée.</p>
-                <a href="#card" class="btn btn-primary" id="button_to_card" div="app">Play</a>
-                </div>
-            </div>
-            <div class="card square-card">
-                <!-- <img src="..." class="card-img-top" alt="..."> -->
-                <div class="card-body">
-                <h5 class="card-title">Cyberpong</h5>
-                <p class="card-text">Cyberpong 3D est une version 3D du célèbre jeu Pong.</p>
-                <a href="#pong3D" class="btn btn-primary" id="button_to_3d" div="app">Play</a>
-                </div>
-            </div>
-        </div>
-    </div>
+	<div class="grid-container">
+		<div class="grid-item navbar" id = "navbar"></div>
+		<div class="grid-item" id = "main-grid">
+			
+			<div class="card square-card">
+				<!-- <img src="..." class="card-img-top" alt="..."> -->
+				<div class="card-body">
+				<h5 class="card-title">RolanGapong</h5>
+				<p class="card-text">RolanGapong est une version revisitée du célèbre jeu Pong, avec des thèmes inspirés des plus grands tournois de tennis.</p>
+				<a href="#pong" class="btn btn-primary" id="button_to_pong" div="app">Play</a>
+				</div>
+			</div>
+			<div class="card square-card">
+				<!-- <img src="..." class="card-img-top" alt="..."> -->
+				<div class="card-body">
+				<h5 class="card-title">Card Game</h5>
+				<p class="card-text">Card Game est un jeu de cartes classique où les joueurs doivent retrouver des paires identiques parmi des cartes disposées face cachée.</p>
+				<a href="#card" class="btn btn-primary" id="button_to_card" div="app">Play</a>
+				</div>
+			</div>
+			<div class="card square-card">
+				<!-- <img src="..." class="card-img-top" alt="..."> -->
+				<div class="card-body">
+				<h5 class="card-title">Cyberpong</h5>
+				<p class="card-text">Cyberpong 3D est une version 3D du célèbre jeu Pong.</p>
+				<a href="#pong3D" class="btn btn-primary" id="button_to_3d" div="app">Play</a>
+				</div>
+			</div>
+		</div>
+	</div>
 </div>
 <!-- <div id="pong-container" style="display: none;"></div>
 <div id="card-container" style="display: none;"></div> -->
