+++ srcs/requirements/apache/volume/pages/navbar.html
@@ -1,0 +1,36 @@
+<div class="grid-item navbar">
+	<nav class="navbar navbar-expand-lg bg-body-tertiary">
+		<div class="container-fluid">
+		  <a class="navbar-brand" href="#home" div="app">Transcendance</a>
+		  <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarSupportedContent" aria-controls="navbarSupportedContent" aria-expanded="false" aria-label="Toggle navigation">
+			<span class="navbar-toggler-icon"></span>
+		  </button>
+		  <div class="collapse navbar-collapse" id="navbarSupportedContent">
+			<ul class="navbar-nav me-auto mb-2 mb-lg-0">
+			  <li class="nav-item">
+				<a href="#home" div="app" class="nav-link active" id="home_button" aria-current="page" >Home</a>
+			  </li>
+			  <li class="nav-item">
+				<a href="#charts" div="app" class="nav-link" >My Stats</a>
+			  </li>
+			  <li class="nav-item dropdown">
+				<a class="nav-link dropdown-toggle" href="#home" div="app" role="button" data-bs-toggle="dropdown" aria-expanded="false">
+				  Games
+				</a>
+				<ul class="dropdown-menu">
+				  <li><a class="dropdown-item" href="#pong" div="app">RolanGapong</a></li>
+				  <li><a class="dropdown-item" href="#card" div="app">Card game</a></li>
+				  <li><hr class="dropdown-divider"></li>
+				  <li><a class="dropdown-item" href="#pong3D" div="app">Cyberpong</a></li>
+				</ul>
+			  </li>
+			</ul>
+			<div class="d-flex ms-auto">
+				<a href=#login class="btn btn-outline-success me-2" id="button_to_login" div="main-grid">login</a>
+				<a href=#register class="btn btn-outline-success" id="button_to_register" div="main-grid">register</a>
+							  <button class="btn btn-outline-success" id="button_to_42api">42-connector</button>
+			  </div>
+		  </div>
+		</div>
+	</nav>
+</div>
